requirements.txt
langchain==0.0.207
sentence_transformers==2.2.2
gpt4all==0.3.6
chromadb==0.3.29
tiktoken==0.4.0
python-dotenv==1.0.0
fire==0.5.0
flake8==6.0.0
flake8-black==0.3.6